feat: simplify main.ts by movinfg icp handler into seperated file
