+++ src/electron/Ipc.Electron.ts
@@ -1,5 +1,5 @@
 import { BrowserWindow, ipcMain, WebFrameMain } from "electron";
-import { getUiPath } from "./pathResolver.js";
+import { getUiPath } from "./utils/pathResolver.js";
 import { isDev } from "./utils/devUtils.js";
 import { pathToFileURL } from "url";
 import { logger } from "./utils/logger.js";
